Update play.cpp
The problem of the delay while inserting the random 2 values has been addressed. This was done by using a set in order to store the elements of the board which were empty and randomly selecting any one element from the set. Earlier the position in which the 2 had to be inserted was randomly selected from all 16 possible values which had led to a delay when most of the board had already been filled.